Fixed Create Report Inconsistencies
Task: "4. "Create Report" button looks different for "Dashboard" and "All Reports" screen. There is a "+" in front of the "Create Report" text for the "All Report" screen button and it is not bolded like the "Dashboard" screen button. Make the "Create Report" button the same text and formatting for both screens."